# cmd/web/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing romantic events in SimpEngine project",
        "title": "SimpEngine API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/romantic-event": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Views a RomanticEvents of current User.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "romantic_event"
                ],
                "summary": "View RomanticEvents by User",
                "responses": {
                    "200": {
                        "description": "List of RomanticEvents",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.RomanticEvent"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates an existing RomanticEvent by ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "romantic_event"
                ],
                "summary": "Updates romantic event",
                "parameters": [
                    {
                        "description": "Romantic event data",
                        "name": "romanticEvent",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/romanticevent.RomanticEventRequestDTO"
                        }
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Romantic event id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid request (invalid ID, invalid body, or validation error)",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized (invalid credentials)",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Simp target/Romantic event not found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Cannot edit event with given status",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates romantic event by user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "romantic_event"
                ],
                "summary": "Creates romantic event",
                "parameters": [
                    {
                        "description": "Romantic event data",
                        "name": "romanticEvent",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/romanticevent.RomanticEventRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/romanticevent.RomanticEventResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid request (invalid ID, invalid body, or validation error)",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized (invalid credentials)",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Simp target not found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/romantic-event/{event_id}/steps": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add event step to romantic event. Only Romantic event owner, can add steps to it.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "romantic_event"
                ],
                "summary": "Add event step to romantic event",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Romantic event id",
                        "name": "event_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Event step data",
                        "name": "eventStep",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/romanticevent.RomanticEventRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/romanticevent.EventStepResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid request (invalid ID, invalid body, or validation error)",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized (invalid credentials)",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Romantic event not found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/romantic-event/{event_id}/steps/{step_id}/options": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add option to romantic event step. Only Romantic event owner, can add option to event step.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "romantic_event"
                ],
                "summary": "Add option to event step",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Romantic event id",
                        "name": "event_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Event step id",
                        "name": "step_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Step option data",
                        "name": "stepOption",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/romanticevent.StepOptionRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Step option data",
                        "schema": {
                            "$ref": "#/definitions/romanticevent.StepOptionResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid request (invalid ID, invalid body, or validation error)",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized (invalid credentials)",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Romantic event/Romantic event step not found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "INVALID_BODY"
                },
                "message": {
                    "type": "string",
                    "example": "Invalid request"
                }
            }
        },
...
        "model.RomanticEvent": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "event_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "public_token": {
                    "type": "string"
                },
                "published_at": {
                    "type": "string"
                },
                "simp_target_id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/model.RomanticEventStatus"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "romanticevent.EventStepResponseDTO": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "event_order": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "romanticevent.RomanticEventRequestDTO": {
            "type": "object",
            "required": [
                "description",
                "event_date",
                "simp_target_id",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "event_date": {
                    "type": "string"
                },
                "simp_target_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string",
                    "minLength": 3
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}